ldr的offset超过范围时，直接将偏移量存入寄存器，不取绝对值。添加错误码，定位错误。

+++ src/compiler/asm/converter/LAddrConverter.java
@@ -35,7 +35,7 @@ public class LAddrConverter extends AsmConverter{
                 builder.add(rd,rd,new ImmOperand(retIr.offset.item*ConstDef.WORD_SIZE));
             }
         }else if(symbol instanceof ParamSymbol){ //todo 这么处理好吗
-            long offsetFP = AsmUtil.getSymbolOffsetFp(symbol);
+            int offsetFP = AsmUtil.getSymbolOffsetFp(symbol);
             if(retIr.offset.isData)
             {
                 offsetFP+=(retIr.offset.item*ConstDef.WORD_SIZE);
